Add Toltec logo
This PR adds the vector files for the new Toltec logo as discussed in #18.
The logo comes in three versions, with various levels of detail suitable
for rendering to different sizes:

* Tiny, used mostly for favicons or otherwise very small icons (16px - 32px)
* Small, used in most cases (64px - 128px)
* Large, intended for use on the home page of the website (256px and larger)

The logo has been designed to work both for dark and light backgrounds.
The source files in this PR have not been optimized, this is to make their later editing easier.
The files should be optimized when shipped to the website.